Create new file for trajectory streamer
