+++ gelatio_boy_control/CMakeLists.txt
@@ -133,7 +133,7 @@ include_directories(
 ## Declare a C++ library
  add_library(${PROJECT_NAME}
   src/hand_controller.cpp
- )
+         include/gelataio_boy_control/plan_executor.h)
 
 ## Add cmake target dependencies of the library
 ## as an example, code may need to be generated before libraries
@@ -143,7 +143,7 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(scooping_planning src/scooping_planning.cpp)
+add_executable(scooping_planning src/scooping_planning.cpp src/plan_executor.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
